refac: 회원가입 로직 분리
