+++ src/types/auth.ts
@@ -16,3 +16,16 @@ export interface SignupRequest {
   profileImage?: string;
   currentLocation?: string;
 }
+
+export type GenderStrict = 'MALE' | 'FEMALE';
+export type Gender = '' | GenderStrict;
+
+export interface Terms {
+  all: boolean;
+  privacy: boolean;
+  age: boolean;
+  marketing: boolean;
+}
+
+export const isGenderStrict = (g: Gender): g is GenderStrict =>
+  g === 'MALE' || g === 'FEMALE';
